requirements.txt
# Dependencias principales
flask==2.3.3
flask-cors==4.0.0
selenium==4.15.2
pandas==2.1.3
openpyxl==3.1.2
pymysql==1.1.0
google-generativeai==0.3.0
requests==2.31.0
PyPDF2==3.0.1
PyMuPDF==1.23.8
webdriver-manager==4.0.1
cryptography==45.0.5
cloud-sql-python-connector[pymysql]

# Dependencias adicionales
python-dotenv==1.0.0
werkzeug==2.3.7
